Develop

+++ src/components/Todo.js
@@ -1,5 +1,4 @@
 import React, { useState, useEffect } from 'react';
-import { MdAdd as AddButtonIcon } from 'react-icons/md';
 import styled from 'styled-components';
 import SubjectForm from './SubjectForm';
 import { TODO_LIST_KEY } from '../common/constants/index';
@@ -21,9 +20,31 @@ const Todo = ({ toggleTheme }) => {
       window.alert('과목명에 내용을 입력 해주세요 ');
       return;
     }
+    if (!subject.end_at) {
+      window.alert('마감일을 선택 해주세요 ');
+      return;
+    }
 
     setTodos(prevTodos => {
       const newTodos = [...prevTodos, subject];
+      localStorage.setItem(TODO_LIST_KEY, JSON.stringify(newTodos));
+      return newTodos;
+    });
+  };
+
+  const checkSubjectHandler = ({ subjectId }) => {
+    const isContinue = window.confirm('완료로 변경 하시겠어요?');
+
+    if (!isContinue) return;
+
+    setTodos(prev => {
+      const newTodos = prev.map(item => {
+        if (item.id === subjectId) {
+          return { ...item, completed: true };
+        }
+        return item;
+      });
+
       localStorage.setItem(TODO_LIST_KEY, JSON.stringify(newTodos));
       return newTodos;
     });
@@ -41,6 +62,34 @@ const Todo = ({ toggleTheme }) => {
     });
   };
 
+  const editSpaceIsVisibleHandler = ({ subjectId }) => {
+    setTodos(
+      todos.map(item => {
+        if (item.id === subjectId) {
+          return { ...item, editSpace: true };
+        }
+        return item;
+      }),
+    );
+  };
+
+  const addDetailHandler = ({ todoItem }) => {
+    const val = window.prompt('메모를 작성 해주세요');
+    if (!val) {
+      alert('내용을 입력 해주세요');
+      return;
+    }
+    setTodos(prev => {
+      const newTodo = prev.map(item =>
+        item.id === todoItem.id
+          ? { ...item, children: [...item.children, val] }
+          : item,
+      );
+      localStorage.setItem(TODO_LIST_KEY, JSON.stringify(newTodo));
+      return newTodo;
+    });
+  };
+
   return (
     <TodoConatiner>
       <Toggler toggleTheme={toggleTheme} />
@@ -48,7 +97,11 @@ const Todo = ({ toggleTheme }) => {
       <TodoList
         todos={todos}
         setTodos={setTodos}
+        editSpaceIsVisibleHandler={editSpaceIsVisibleHandler}
+        addSubjectHanlder={addSubjectHanlder}
         removeSubjectHandler={removeSubjectHandler}
+        addDetailHandler={addDetailHandler}
+        checkSubjectHandler={checkSubjectHandler}
       />
     </TodoConatiner>
   );
@@ -56,34 +109,6 @@ const Todo = ({ toggleTheme }) => {
 
 const TodoConatiner = styled.div`
   width: 100%;
-`;
-
-const TodoFormWrapper = styled.form`
-  display: flex;
-  background-color: ${({ theme }) => theme.backgroundColor};
-`;
-
-const Input = styled.input`
-  padding: 8px;
-  font-size: 18px;
-  line-height: 1.5;
-  color: black;
-  flex: 1;
-`;
-
-const InsertButton = styled.div`
-  color: ${({ theme }) => theme.textColor};
-  padding-left: 16px;
-  padding-right: 16px;
-  font-size: 25px;
-  display: flex;
-  align-items: center;
-  cursor: pointer;
-  transition: 0.1s background ease-in;
-  background-color: ${({ theme }) => theme.primaryColor};
-  &:hover {
-    background - color: ${({ theme }) => theme.primaryTintColor};
-  }
 `;
 
 export default Todo;
